Bug #215225 (Part only)
Made a first pass at this.

The src/external folder is now empty, and all the "Iext" classes have been
removed, with their methods and fields migrated to the simple "I" versions.

Some restructuring of the packages has been completed as well - so we now have
model.artifacts instead of artifacts.model for example.

More work is required:

Remove and methods with Iext in their name for example (I got most of them but
I'm sure I missed a few.)

More re-structuring of the packages.

Review of the classes for sensible/consistent methods

Javadoc updates


I have tested the basic code by creating a new Project with a few simple
entities, creating a profile and deploying that, and then making a simple
plugin that I ran against the simple Project.

I was also able to browse existing projects with no problems.