Bug 222018 & 222517
Attributes, Literals and Methods order can now be defined by the user  wuing:

Add
Up
Down
Remove

The tables can also be sorted by the name (up/down), and in the case of literals  by the value.

Pressing up/down disables any sorting, and added elements will be after the current position or in the defined sort position if any sort is applied.

The explorer is updated on a save to match the above order.