Bug 222336.
When dragging components, the information other than that held i the diagram for display was being list.

The Attribute, Literal and Method on the diagram now have an IFiled, ILietral andIMethod thats holds a reference to the related item, so that the "new" itme can be populated with that when the drop occurs.

Updates have been amde to keep the model and diagram  in synch.