Bug 224666
The UML export now does the following :

creates a separate UML module for each Dependency and Reference Project, in addition to the actual project itself. These are all placed in the same folder, alongside the profile etc.

This still has some shortcomings:

a) distrib.core.model.uml is ALWAYS generated.
b) Dependency of Dependency causes issues ( eg If my project references a project that has a module dependency, the export tries to resolve things in the remote project and can't unless the dependency is also made on my project).
c) the xmi id are recreated at every export - I don't know if this is a real problem when using the import in other tools, but I suspect so.