Updated the logic a bit.
Now will only add an Association if both ends are in scope. This meant that the Association  handling has moved "up" a bit, so we can check the ends BEFORE the assoc gets added - instead of after.

I also generate a "warning" in the GUI if this situation occurs when the Facet is applied there.