Bug 251415
As per the bug, the value being found was not being allocated to a variable.

Clearly an error, but I have no idea how to test this as I can't work out the effect of the bug - all the cases I tried had "valid" entries, so never got into this bit!