Bug 235883
Couple of changes :

Removed a check that was looking for a valid ProjectDetails.NAme - no longer a sensible check!

Override the getName() method for PLuginProjectHanldes as the method did not cope with getting the name of a DEPLOYED plugin.

Updated the getID()method on the Plugin descriptor to use the name instead of ProjectLabel.