Improvements for intermittent failures
Sometimes Tigerstripe would report an incorrect number of configured
generation plugins configured on a Model project. These changes add
enhanced logging to help debug any scenario in the future where an
intermittent error may occure, as well as ensure that code that
references a Tigerstripe Project use APIs to re-use the cached
ITigerstripeModelProject instance, instead of using API's that may
create a new instance in some scenarios, which I believe was causing
some aspects of generation to have either a non-loaded object, or a
disposed object.